+++ packages/koni_jxl/lib/koni_jxl.dart
@@ -6,4 +6,5 @@ export 'src/encode/encoder.dart' show JxlEncoder;
 export 'src/exceptions.dart';
 export 'src/jxl_image.dart' show JxlAnimation, JxlImage;
 export 'src/jxl_info.dart' show JxlInfo;
+export 'src/jxl_limits.dart' show JxlLimits;
 export 'src/streaming.dart' show JxlStreamState, JxlStreamingDecoder;